+++ python/array-generalise.py
