+++ python/simplegame.py
