+++ ruby/lowerAndUpper.rb
